refactor: make separate types
